move UIMF reader into its own module and add attribution
